requirements.txt
streamlit==1.28.0
pandas==2.0.3
plotly==5.17.0
scikit-learn==1.3.0
SpeechRecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.13
numpy==1.24.3
sqlalchemy==2.0.23
pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1
python-dateutil==2.8.2